feat: move msw setup for weather and qotd into test folder
