+++ src/slash-commands/quote-of-the-day/index.test.ts
@@ -1,4 +1,3 @@
-import { faker } from '@faker-js/faker';
 import type { ChatInputCommandInteraction } from 'discord.js';
 import { http, HttpResponse } from 'msw';
 import { beforeEach, describe, expect, it } from 'vitest';
@@ -37,17 +36,6 @@ describe('Get quote of the day test', () => {
   });
 
   it('Should reply with a random quote', async () => {
-    const fakeQuote = faker.lorem.words(25);
-    const sampleQuote = {
-      q: fakeQuote,
-      a: 'Author',
-      h: `<h1>${fakeQuote}</h1>`,
-    };
-    const endpoint = http.get(ZEN_QUOTES_URL, () => {
-      return HttpResponse.json([sampleQuote]);
-    });
-    server.use(endpoint);
-
     await getQuoteOfTheDay(mockInteraction);
     expect(mockInteraction.editReply).toHaveBeenCalledOnce();
   });
